Close F-19: split pipeline.py seal + diagram into leaf modules
Move the F-15 semantic digest into pipeline_seal.py (artifact_digest)
and Mermaid rendering into pipeline_diagram.py (build_mermaid_diagram);
SkyulfPipeline.fingerprint()/to_mermaid() become one-line delegates.
Both are leaf modules with no import-cycle risk; behavior unchanged.
build_mermaid_diagram widens to Sequence/Mapping so callers can pass
PreprocessingStepConfig/ModelConfig TypedDicts without a ty error.
