+++ skyulf-core/tests/unit/test_pipeline_coverage.py
@@ -1,9 +1,10 @@
 """Coverage-focused tests for skyulf.pipeline (the top-level SkyulfPipeline).
 
 Exercises branches not hit by the existing end-to-end test in test_pipeline.py:
-* `_artifact_digest` semantic (F-15): determinism, weight sensitivity,
-  tree structures, tuned ``(model, TuningResult)`` tuples, and the
-  fail-loud ``TypeError`` that replaced the old ``repr`` fallback.
+* `artifact_digest` (moved to `skyulf.pipeline_seal` in the F-19 split): the
+  F-15 semantic digest's determinism, weight sensitivity, tree structures,
+  tuned ``(model, TuningResult)`` tuples, and the fail-loud ``TypeError``
+  that replaced the old ``repr`` fallback.
 * `_init_model_estimator` early return, registry-only model resolution
   (including the `hyperparameter_tuner` base-model branch), and the
   "unknown model type" / "partially registered" error paths.
@@ -22,11 +23,12 @@ from sklearn.linear_model import LogisticRegression
 
 from skyulf.modeling._tuning.engine import TuningApplier, TuningCalculator
 from skyulf.modeling._tuning.schemas import TuningResult
-from skyulf.pipeline import SkyulfPipeline, _artifact_digest
+from skyulf.pipeline import SkyulfPipeline
+from skyulf.pipeline_seal import artifact_digest
 from skyulf.registry import NodeRegistry
 
 # ---------------------------------------------------------------------------
-# _artifact_digest
+# artifact_digest (skyulf.pipeline_seal)
 # ---------------------------------------------------------------------------
 
 
@@ -40,7 +42,7 @@ def _small_xy() -> tuple[np.ndarray, np.ndarray]:
 def test_artifact_digest_is_deterministic_for_same_estimator() -> None:
     X, y = _small_xy()
     est = LogisticRegression(random_state=42).fit(X, y)
-    assert _artifact_digest(est) == _artifact_digest(est)
+    assert artifact_digest(est) == artifact_digest(est)
 
 
 def test_artifact_digest_detects_different_weights_same_hyperparams() -> None:
@@ -51,7 +53,7 @@ def test_artifact_digest_detects_different_weights_same_hyperparams() -> None:
     flipped = LogisticRegression(random_state=42).fit(X, 1 - y)
     assert same.get_params() == flipped.get_params()
     assert not (same.coef_ == flipped.coef_).all()
-    assert _artifact_digest(same) != _artifact_digest(flipped)
+    assert artifact_digest(same) != artifact_digest(flipped)
 
 
 def test_artifact_digest_raises_on_undigestible_object() -> None:
@@ -63,16 +65,16 @@ def test_artifact_digest_raises_on_undigestible_object() -> None:
             self.gen = (i for i in range(3))
 
     with pytest.raises(TypeError):
-        _artifact_digest(_Opaque())
+        artifact_digest(_Opaque())
 
 
 def test_artifact_digest_is_key_order_insensitive_for_dicts() -> None:
     """Preprocessing artifacts are config dicts; insertion order must not
     matter for the seal."""
-    assert _artifact_digest({"a": 1, "b": 2.0, "c": [1, 2]}) == _artifact_digest(
+    assert artifact_digest({"a": 1, "b": 2.0, "c": [1, 2]}) == artifact_digest(
         {"c": [1, 2], "b": 2.0, "a": 1}
     )
-    assert _artifact_digest({"a": 1}) != _artifact_digest({"a": 2})
+    assert artifact_digest({"a": 1}) != artifact_digest({"a": 2})
 
 
 def test_artifact_digest_covers_tree_structures() -> None:
@@ -80,9 +82,9 @@ def test_artifact_digest_covers_tree_structures() -> None:
     must walk their node arrays."""
     X, y = _small_xy()
     rf = RandomForestClassifier(n_estimators=3, random_state=42).fit(X, y)
-    assert _artifact_digest(rf) == _artifact_digest(rf)
+    assert artifact_digest(rf) == artifact_digest(rf)
     other = RandomForestClassifier(n_estimators=3, random_state=43).fit(X, y)
-    assert _artifact_digest(rf) != _artifact_digest(other)
+    assert artifact_digest(rf) != artifact_digest(other)
 
 
 def test_artifact_digest_covers_tuned_model_tuples() -> None:
@@ -91,9 +93,9 @@ def test_artifact_digest_covers_tuned_model_tuples() -> None:
     X, y = _small_xy()
     est = LogisticRegression(random_state=42).fit(X, y)
     result = TuningResult(best_params={"C": 1.0}, best_score=0.9, n_trials=1, trials=[])
-    assert _artifact_digest((est, result)) == _artifact_digest((est, result))
+    assert artifact_digest((est, result)) == artifact_digest((est, result))
     changed = TuningResult(best_params={"C": 10.0}, best_score=0.9, n_trials=1, trials=[])
-    assert _artifact_digest((est, result)) != _artifact_digest((est, changed))
+    assert artifact_digest((est, result)) != artifact_digest((est, changed))
 
 
 def test_fingerprint_of_tree_pipeline_is_deterministic_and_data_sensitive() -> None:
